+++ packages/solid-hotkeys-devtools/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/solid-hotkeys-devtools
 
+## 0.3.0
+
+### Minor Changes
+
+- feat: overhaul sequence-manager and hooks to be in feature parity with hotkey-manager. ([#21](https://github.com/TanStack/hotkeys/pull/21))
+
+### Patch Changes
+
+- Updated dependencies [[`7328e36`](https://github.com/TanStack/hotkeys/commit/7328e360f0e99b5374fb97c07e0f2a500d8e5b9c)]:
+  - @tanstack/hotkeys-devtools@0.3.0
+
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies [[`8ae6b64`](https://github.com/TanStack/hotkeys/commit/8ae6b64ef10b53186c367f594f81ecdec15071d7)]:
+  - @tanstack/hotkeys-devtools@0.2.0
+
 ## 0.1.3
 
 ### Patch Changes